Update triangular optimization (#77)
- Make optimization scale and shift invariant
- Instead of optimizing in CDF space, we optimize along the real line (but after scaling the problem)
- Check closeness of distributions using point wise RMSE between PDFs

The big trick to was to move away from comparing distances in CDF-space (numbers between 0 and 1) because in that space there is no feedback for the optimizer when going outside of the bounds - evaluating CDF of a triangular a little bit outside or far outside gives the same result.

Another problem with the existing routine was that the fsolve optimizer would not obey minimum < mode < maximum. By re-parametrizing the search space we can avoid this issue too.